requirements.txt
channels[daphne]
channels-redis
djangorestframework
django-cors-headers